move out the constant fixed base point
